backend/ReceivePubSub/requirements.txt
requests
boto3
google-cloud-pubsub